Few simplifications.

+++ Agent/Sheep/PassiveSheep.cs
@@ -2,7 +2,7 @@
 {
     class PassiveSheep : Sheep
     {
-        public PassiveSheep(float x, float y) : base(x, y)
+        public PassiveSheep(double x, double y) : base(x, y)
         {
         }
     }
