+++ Simulation/FitnessCounters/IFitnessCounter.cs
@@ -4,6 +4,6 @@ namespace Simulations
 {
     public interface IFitnessCounter
     {
-        float CountFitness(Team team);
+        double CountFitness(Team team);
     }
 }
